Stylework
Edited Files: TestStyles.css, About.html, browse.hmtl

TestStyles.css- Added the classes: category, cat-header, cat-hot, and nav-tab-active.  Also fixed some small margin errors with-in other classes

About.hmtl: Added the re-styled navbar to this page

browse.html: Styled the "buttons" that will lead the user to the category pages, added the restyled navbar to this page, made sure that the categories section displayed properly and had room for a description to be added if the need arises.